Add convenience functions to test for equality
assert_equal() and assert_identical() allow for testing for loose and
strict equality respectively, since these are expected to be the most
common cases. These functions mitigate some of the shortcomings of
using assert() directly, specifically where the context might be overly
noisy, such as when referencing an object property or an array element,
or missing desired information, such as when referencing $this or the
result of a function call. Instead of assigning results to temporary
variables for use in assert(), they can be used directly as arguments
to these functions.