Buffer and report output during test runs
Output buffering is done in the Reporter. A healthy discussion can
probably be had about whether it's appropriate for this to be the
Reporter's responsibility, but it's a convenient, centralized place to
handle it, and the overhead of adding an additional dependency doesn't
currently seem justified.

Of note, tests methods are executed in a common buffer with their setup
and teardown methods. This seems to be the right thing to do, as it
allows the buffer to be managed as part of the test fixture. The
TestReporter provides an example of this: the report header is output
in setup(), combined with output in the individual tests and checked,
and then cleared in teardown() to prevent it from showing up in the
test report. This process would be slightly more tedious if each method
was executed in its own buffer.

As a consequence, failures and errors can't be reported immediately,
but need to be held until the end of the test. One nice side effect of
this is that setup and teardown can now be associated with the specific
test they were run for.