Report events in the order in which they occur
There doesn't seem to be a compelling reason to order and report events
by their type, so events are now collected and then reported in the
order in which they occur. One example of the benefits of this is that
output is now displayed in proximity with other events that might be
generated from the same test, such as a failure or an error.

The format of the report has also changed rather dramatically as a
result. The new format now emulates that of the Python unittest module.