Only report output for non-passing tests
The theory behind this being that the user will usually emit output to
troubleshoot an error or a failing test. Consequently, suppressing
output for passed tests reduces the noise in the test report.

Output is still signalled in the real-time progress display (by
printing an 'O') and included in the total output count. This will
remind/inform the user that their program is emitting output even if
all tests are passing.