Update TODO.txt
And ... it worked on my computer and not on Travis because easytest runs
in 'verbose' mode on Travis, so converting the skipped tests, which are
only reported in 'verbose' mode, to a string in order to generate the
test report caused the ob_flush that had been added for debugging to be
called without an output buffer in place. Mystery solved!